001_first_session/pyproject.toml
[project]
name = "firstsession"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "lgcns-ai-keonho-kim", email = "kkh93@lgcns.com" }
]
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "google-generativeai>=0.8.6",
    "httpx>=0.28.1",
    "json-repair>=0.55.1",
    "langchain>=1.2.7",
    "langchain-google-genai>=4.2.0",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.7",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]

[project.scripts]
firstsession = "firstsession:main"

[build-system]
requires = ["uv_build>=0.8.19,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
]
